extract POSCollectOrderPaymentAnalytics to file
